+++ CHANGES.rst
@@ -2,6 +2,7 @@
 -----
 * Base(Scoped)?(Id)?NameMixin now disallows blank names by default. Bumped version
   number since this is a non-breaking incompatible change.
+* JsonDict now uses JSONB on PostgreSQL 9.4.
 
 0.4.3
 -----